A little improvement on memory consumption. The main bottleneck is a GC: - for scale=14 mutator works 46% of time, - for scale=16 mutator works 37% of time and - for scale=18 it works 28% of time.
I will stay with GHC heap structures as long as possible as
they are easy to use in parallel setting. "As long as possible"
is about two weeks. Then I either made it or I will switch to
modifable structures in allocated memory.

My priority right now is to shrink memory consumption.

Change-Id: I71dc521f7cba7e2b58b778b5f742887108e84f16
Reviewed-on: http://review.assembla.com/56043
Reviewed-by: kyrill.briantsev <kyrill.briantsev@parsci.com>
Tested-by: kyrill.briantsev <kyrill.briantsev@parsci.com>
Reviewed-by: Rich Lyman <richard.lyman@parsci.com>
Tested-by: Rich Lyman <richard.lyman@parsci.com>